修改选中供应商逻辑；

+++ StorageApp/Business/Main/Model/DSSSupplierModel.swift
@@ -10,7 +10,7 @@ import Foundation
 import ObjectMapper
 
 class DSSSupplierModel: NSObject, Mappable {
-    var itemID          : Int64!
+    var itemID          : Int!
     var name            : String!
     var isSelected      : Bool!
